+++ datatypes/src/jvmTest/kotlin/Asn1EncodingTest.kt
@@ -21,6 +21,50 @@ class Asn1EncodingTest : FreeSpec({
         fromBitSet.encodeToTlv().toDerHexString() shouldBe "0304066E5DC0"
         fromBitSet.toBitSet().toBitString() shouldBe "011011100101110111"
         fromBitSet.toBitSet() shouldBe bitSet
+
+        Asn1BitString.decodeFromTlv(asn1Sequence { bitString(bitSet) }.children.first() as Asn1Primitive)
+            .toBitSet() shouldBe bitSet
+    }
+
+    "OCTET STRING Test" {
+        val seq = asn1Sequence {
+            octetStringEncapsulated {
+                sequence { utf8String("foo") }
+                set { utf8String("bar") }
+                append(Asn1String.Printable("a").encodeToTlv())
+            }
+            octetString(byteArrayOf(17))
+
+
+
+            octetString(
+                asn1Set {
+                    int(99)
+                    octetString(byteArrayOf(1, 2, 3))
+                    octetStringEncapsulated {
+                        append(
+                            Asn1EncapsulatingOctetString(
+                                listOf(
+                                    Asn1PrimitiveOctetString(
+                                        byteArrayOf(
+                                            7,
+                                            6,
+                                            3,
+                                        )
+                                    )
+                                )
+                            )
+                        )
+                    }
+                }.derEncoded
+            )
+            tagged(9u) { append(Clock.System.now().encodeToAsn1UtcTime()) }
+            octetString(byteArrayOf(17, -43, 23, -12, 8, 65, 90))
+            bool(false)
+            bool(true)
+        }
+        val parsed = Asn1Element.parse(seq.derEncoded)
+        println(parsed.prettyPrint())
     }
 
     "Ans1 Number encoding" - {
@@ -39,7 +83,7 @@ class Asn1EncodingTest : FreeSpec({
 
         "longs" - {
             checkAll(iterations = 15000, Arb.long()) {
-                val seq = asn1Sequence { long { it } }
+                val seq = asn1Sequence { long(it) }
                 val decoded = (seq.nextChild() as Asn1Primitive).readLong()
                 decoded shouldBe it
             }
@@ -47,7 +91,7 @@ class Asn1EncodingTest : FreeSpec({
 
         "ints" - {
             checkAll(iterations = 15000, Arb.int()) {
-                val seq = asn1Sequence { int { it } }
+                val seq = asn1Sequence { int(it) }
                 val decoded = (seq.nextChild() as Asn1Primitive).readInt()
                 decoded shouldBe it
             }
@@ -62,45 +106,45 @@ class Asn1EncodingTest : FreeSpec({
         tree.derEncoded shouldBe certBytes
     }
 
-    "Encoding and parsing results in the same bytes" {
+    "Old and new encoder produce the same bytes" {
 
         val instant = Clock.System.now()
 
         val sequence = asn1Sequence {
-            tagged(31u) {
-                Asn1Primitive(BERTags.BOOLEAN, byteArrayOf(0x00))
+            tagged(1u) {
+                append(Asn1Primitive(BERTags.BOOLEAN, byteArrayOf(0x00)))
             }
             set {
                 sequence {
                     setOf {
-                        printableString { "World" }
-                        printableString { "Hello" }
+                        printableString("World")
+                        printableString("Hello")
                     }
                     set {
-                        printableString { "World" }
-                        printableString { "Hello" }
-                        utf8String { "!!!" }
+                        printableString("World")
+                        printableString("Hello")
+                        utf8String("!!!")
                     }
 
                 }
             }
             asn1null()
 
-            oid { ObjectIdentifier("1.2.603.624.97") }
+            append(ObjectIdentifier("1.2.603.624.97"))
 
-            utf8String { "Foo" }
-            printableString { "Bar" }
+            utf8String("Foo")
+            printableString("Bar")
 
             set {
-                int { 3 }
-                long { -65789876543L }
-                bool { false }
-                bool { true }
+                int(3)
+                long(-65789876543L)
+                bool(false)
+                bool(true)
             }
             sequence {
                 asn1null()
-                string { Asn1String.Numeric("12345") }
-                utcTime { instant }
+                append(Asn1String.Numeric("12345"))
+                utcTime(instant)
             }
         }