+++ CHANGELOG.md
@@ -12,3 +12,7 @@
  * Bugfixes and streamlining all over the place
  * Proper BIT STRING
  * BitSet (100% Kotlin BitSet implementation)
+ * Recursively parsing (and encapsulating) ASN.1 structures in OCTET Strings
+ * Initial pretty-printing of ASN.1 Strucutres
+ * Massive ASN.1 builder DSL streamlining
+ * More convenient explicit tagging